requirements.txt
django
djangorestframework
djoser
corsheaders